+++ src/components/Themeroller/ItemValue/ColorValue/ColorValue.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { bem } from '../../../../utils';
 import { Input } from '../../../Input';
 import { Text } from '../../../Text';
-import { ThemeColorItem } from '../../themeConfigTypes';
+import { ThemeColorItem } from '../../../../themes/themerollerConfig';
 import styles from './ColorValue.scss';
 
 type Props = {
