+++ src/components/Themeroller/ItemValue/NumberValue/NumberValue.tsx
@@ -1,0 +1,22 @@
+import * as React from 'react';
+import { Input } from '../../../Input';
+import { ThemeNumberItem } from '../../../../themes/themerollerConfig';
+
+type Props = {
+    item: ThemeNumberItem;
+    onChange: (value: ThemeNumberItem) => void;
+};
+
+export const NumberValue: React.FC<Props> = ({ item, onChange }) => {
+    const handleOnChange = (event) => {
+        onChange({
+            ...item,
+            value: event.target.value,
+        });
+    };
+    return <Input type="number" size="small" value={item.value} onChange={handleOnChange} />;
+};
+
+NumberValue.displayName = 'NumberValue';
+
+NumberValue.defaultProps = {};
