+++ src/components/Themeroller/ItemValue/NumberValue/index.ts
@@ -1,0 +1,1 @@
+export { NumberValue } from './NumberValue';
