+++ src/components/Themeroller/index.ts
@@ -1,3 +1,2 @@
 export * from './Themeroller';
 export * from './ThemerollerActions';
-export * from './themeConfigTypes';
