+++ src/themes/themerollerConfig/index.ts
@@ -1,0 +1,2 @@
+export * from './themeConfigType';
+export * from './themeConfig';
